requirements.txt
autopep8==1.6.0
click==8.0.3
colorama==0.4.6
dataclasses==0.6
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-dotenv==0.19.1
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
Werkzeug==2.0.2
